Convert SignUp page to using scss stuff
